+++ packages/genesys-spark-components/jest.config.ts
@@ -17,7 +17,8 @@ const config: Config = {
   }),
   collectCoverage: true,
   coverageDirectory: 'build/test-reports/coverage',
-  coverageReporters: ['json', 'lcov', 'clover'],
+  coverageReporters: ['json', 'lcov', 'clover', 'text'],
+  coveragePathIgnorePatterns: ['/tests/'],
   setupFilesAfterEnv: [
     '<rootDir>/src/test/setupTests.ts',
     '<rootDir>/src/test/setupAxeTests.ts'