+++ packages/genesys-spark-components/src/components/stable/gux-dropdown-multi/gux-dropdown-multi.tsx
@@ -25,7 +25,7 @@ import translationResources from './i18n/en.json';
 import {
   getSearchOption,
   setInitialActiveOption,
-  getOptionDefaultSlot,
+  getOptionDefaultSlotText,
   convertValueToArray
 } from '../gux-listbox/gux-listbox.service';
 import { GuxFilterTypes } from '../gux-dropdown/gux-dropdown.types';
@@ -478,8 +478,7 @@ export class GuxDropdownMulti {
     if (textInputLength > 0 && !this.loading) {
       const option = getSearchOption(this.listboxElement, textInput);
       if (option && this.filterType !== 'custom') {
-        const optionSlotTextContent =
-          getOptionDefaultSlot(option)?.textContent.trim();
+        const optionSlotTextContent = getOptionDefaultSlotText(option);
         return optionSlotTextContent?.substring(textInputLength);
       }
 
@@ -512,7 +511,7 @@ export class GuxDropdownMulti {
   private getSelectedOptionValueString(): string {
     return this.getOptionElementByValue(this.value)
       .map(option => {
-        return getOptionDefaultSlot(option)?.textContent.trim();
+        return getOptionDefaultSlotText(option);
       })
       .join(', ');
   }
@@ -586,7 +585,6 @@ export class GuxDropdownMulti {
                   onFocus={() => (this.expanded = true)}
                   disabled={this.disabled}
                 ></input>
-                {this.renderTargetContent()}
               </div>
             </div>
           </div>
@@ -627,7 +625,7 @@ export class GuxDropdownMulti {
           aria-haspopup="listbox"
           aria-expanded={this.expanded.toString()}
         >
-          {!this.hasTextInput() && this.renderTargetContent()}
+          {this.renderTargetContent()}
           {this.renderTag()}
           {this.renderRadialLoading()}
           <gux-icon
@@ -643,7 +641,7 @@ export class GuxDropdownMulti {
     ) as JSX.Element;
   }
   private renderTargetContent(): JSX.Element {
-    if (!(this.expanded && this.hasTextInput())) {
+    if (!this.hasTextInput()) {
       return (
         <span class="gux-field-content">{this.renderTargetDisplay()}</span>
       ) as JSX.Element;